Refactor:#76 - standardize goal icon naming, location, and format
- Relocate from public/assets/icons/goalIcon.svg to src/assets/icons/goal.svg
- Migrate from next/Image to SVGR import
- Apply consistent SVGR props
