+++ src/components/goals/GoalCardContent.tsx
@@ -3,7 +3,7 @@
 import Image from 'next/image';
 import { useRouter } from 'next/navigation';
 
-import GoalIcon from '@/../public/assets/icons/goalIcon.svg';
+import GoalIcon from '@/assets/icons/goal.svg';
 import { Button } from '@/components/ui/Button';
 import { GoalSummary } from '@/interfaces/goal';
 import { TodoSummary } from '@/interfaces/todo';
@@ -40,7 +40,12 @@ export default function GoalCardContent({
         <div className="flex flex-col gap-20">
           <div className="flex flex-col gap-12">
             <div className="flex items-center gap-8">
-              <GoalIcon className={`h-24 w-24 ${getGoalTextColorClass(goal.color)}`} />
+              <GoalIcon
+                className={getGoalTextColorClass(goal.color)}
+                width={24}
+                height={24}
+                fill="currentColor"
+              />
               <h3 className="text-text-01 text-body-sb-20 max-w-296 truncate">{goal.title}</h3>
             </div>
             <div className="flex items-baseline gap-12">
