+++ src/components/goals/goalList/GoalsHeader.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import GoalIcon from '@/../public/assets/icons/goalIcon.svg';
+import GoalIcon from '@/assets/icons/goal.svg';
 
 interface GoalsHeaderProps {
   totalCount: number;
@@ -12,7 +12,7 @@ const GoalsHeader = ({ totalCount }: GoalsHeaderProps) => {
       <div className="flex items-center justify-between">
         <div>
           <div className="flex items-center gap-8">
-            <GoalIcon className="h-24 w-24" />
+            <GoalIcon className="text-snackbar" width={24} height={24} fill="currentColor" />
             <h1 className="text-display-32 text-text-01">모든 목표</h1>
           </div>
